server/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
Flask==2.3.3
Flask-PyMongo==2.3.0
flask-restx==1.2.0
gunicorn==21.2.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
pymongo==4.5.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3.post1
typing_extensions==4.8.0
Werkzeug==2.3.7
